+++ components/frontend/src/source/Source.js
@@ -6,34 +6,12 @@ import { StringInput } from '../fields/StringInput';
 import { Logo } from '../logos/Logo';
 import { ChangeLog } from '../changelog/ChangeLog';
 import { ItemActionButtons } from '../widgets/Button';
-import { ItemBreadcrumb } from '../widgets/ItemBreadcrumb';
 import { copy_source, delete_source, move_source, set_source_attribute } from '../api/source';
 import { ReadOnlyOrEditable } from '../context/ReadOnly';
-import { get_metric_name, get_subject_name } from '../utils';
+import { metric_options } from '../menu_options';
 
 function select_sources_parameter_keys(changed_fields, source_uuid) {
     return changed_fields ? changed_fields.filter((field) => field.source_uuid === source_uuid).map((field) => field.parameter_key) : []
-}
-
-function metric_options(reports, datamodel, current_metric_uuid) {
-    let options = [];
-    reports.forEach((report) => {
-        Object.values(report.subjects).forEach((subject) => {
-            const subject_name = get_subject_name(subject, datamodel);
-            Object.entries(subject.metrics).forEach(([metric_uuid, metric]) => {
-                const metric_name = get_metric_name(metric, datamodel);
-                options.push({
-                    content: <ItemBreadcrumb report={report.title} subject={subject_name} metric={metric_name} />,
-                    disabled: metric_uuid === current_metric_uuid,
-                    key: metric_uuid,
-                    text: report.title + subject_name + metric_name,
-                    value: metric_uuid
-                })
-            })
-        });
-    });
-    options.sort((a, b) => a.text.localeCompare(b.text));
-    return options;
 }
 
 export function Source(props) {