+++ components/server/src/routes/report.py
@@ -18,6 +18,7 @@ from server_utilities.type import ReportId
 
 
 @bottle.post("/api/v2/report/import")
+@bottle.post("/api/v3/report/import")
 def post_report_import(database: Database):
     """Import a preconfigured report into the database."""
     report = dict(bottle.request.json)
@@ -25,6 +26,7 @@ def post_report_import(database: Database):
 
 
 @bottle.post("/api/v2/report/new")
+@bottle.post("/api/v3/report/new")
 def post_report_new(database: Database):
     """Add a new report."""
     report_uuid = uuid()
@@ -36,6 +38,7 @@ def post_report_new(database: Database):
 
 
 @bottle.post("/api/v2/report/<report_uuid>/copy")
+@bottle.post("/api/v3/report/<report_uuid>/copy")
 def post_report_copy(report_uuid: ReportId, database: Database):
     """Copy a report."""
     data = ReportData(database, report_uuid)
@@ -48,6 +51,7 @@ def post_report_copy(report_uuid: ReportId, database: Database):
 
 
 @bottle.get("/api/v2/report/<report_uuid>/pdf")
+@bottle.get("/api/v3/report/<report_uuid>/pdf")
 def export_report_as_pdf(report_uuid: ReportId):
     """Download the report as pdf."""
     host = os.environ.get("PROXY_HOST", "www")
@@ -62,6 +66,7 @@ def export_report_as_pdf(report_uuid: ReportId):
 
 
 @bottle.delete("/api/v2/report/<report_uuid>")
+@bottle.delete("/api/v3/report/<report_uuid>")
 def delete_report(report_uuid: ReportId, database: Database):
     """Delete a report."""
     data = ReportData(database, report_uuid)
@@ -74,6 +79,7 @@ def delete_report(report_uuid: ReportId, database: Database):
 
 
 @bottle.post("/api/v2/report/<report_uuid>/attribute/<report_attribute>")
+@bottle.post("/api/v3/report/<report_uuid>/attribute/<report_attribute>")
 def post_report_attribute(report_uuid: ReportId, report_attribute: str, database: Database):
     """Set a report attribute."""
     data = ReportData(database, report_uuid)
@@ -90,6 +96,7 @@ def post_report_attribute(report_uuid: ReportId, report_attribute: str, database
 
 
 @bottle.get("/api/v2/tagreport/<tag>")
+@bottle.get("/api/v3/tagreport/<tag>")
 def get_tag_report(tag: str, database: Database):
     """Get a report with all metrics that have the specified tag."""
     date_time = report_date_time()